LSR-1006 move page to separate class
